API Error moved to common
