+++ src/common/api/index.ts
@@ -1,2 +1,3 @@
 export * from './types';
+export * from './customTypes';
 export * from './utils';
