+++ src/common/utils.ts
@@ -1,4 +1,5 @@
 import isString from 'lodash/isString';
+import { isApiError } from './api/customTypes';
 
 export const getErrorMessage = (error: unknown) => {
   if (error instanceof Error) {
@@ -8,4 +9,11 @@ export const getErrorMessage = (error: unknown) => {
     return error;
   }
   return 'Unexpected error';
+};
+
+export const getApiErrorMessage = (error: unknown): string => {
+  if (isApiError(error)) {
+    return error.response?.data?.message || error.response?.data.reason || error.message;
+  }
+  return getErrorMessage(error);
 };
