+++ src/ocm/components/clusterDetail/CancelInstallationModal.tsx
@@ -8,11 +8,12 @@ import {
   Text,
   TextContent,
 } from '@patternfly/react-core';
-import { getApiErrorMessage, handleApiError } from '../../api/utils';
-import { updateCluster } from '../../reducers/clusters/currentClusterSlice';
+import { handleApiError } from '../../api';
+import { updateCluster } from '../../reducers/clusters';
 import { useModalDialogsContext } from '../hosts/ModalDialogsContext';
 import { ErrorState, LoadingState } from '../../../common';
 import { ClustersAPI } from '../../services/apis';
+import { getApiErrorMessage } from '../../../common/utils';
 
 const CancelInstallationModal: React.FC = () => {
   const dispatch = useDispatch();
